refactor(shared): move vertex provider metadata to module
